+++ modules/sagemaker/outputs.tf
@@ -1,6 +1,6 @@
 output "notebook_instance" {
   description = "Sagemaker notebook instance Terraform object"
-  value       = aws_sagemaker_notebook_instance.notebook
+  value       = aws_sagemaker_notebook_instance.training_notebook
 }
 
 output "role" {